--- classes/index.js
@@ -1,25 +1,0 @@
-/**
- * node-vt
- * Copyright (C)  2020 Logan Devine
- * 
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-// Import / Export
-module.exports.Err = require('./Err.js');
-module.exports.URLParams = require('./URLParams.js');
-module.exports.Response = require('./Response.js');
-module.exports.Request = require('./Request.js');
-module.exports.APIRequest = require('./APIRequest.js');
-module.exports.Client = require('./Client.js');
